Move emulator to dedicated package and pass backend to command handlers
