+++ emulator/errors.go
@@ -1,0 +1,13 @@
+package emulator
+
+import "golang.org/x/xerrors"
+
+var (
+	// ErrErrUnsupportedRequest is returned by backends when they cannot
+	// process a particular mongo client request.
+	ErrUnsupportedRequest = xerrors.New("unsupported request")
+
+	// ErrInvalidCursor is returned by backends when a request includes an
+	// unknown/invalid cursor ID.
+	ErrInvalidCursor = xerrors.New("invalid cursor")
+)
